api/requirements.txt
anyio==4.8.0
asgiref==3.8.1
bcrypt==4.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cytoolz==1.0.1
diff-match-patch==20241021
Django==5.1.5
django-import-export==4.3.4
django-unfold==0.46.0
dnspython==2.7.0
eth-abi==4.2.1
eth-hash==0.7.1
eth-typing==5.1.0
eth-utils==5.2.0
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
import-export==0.3.1
iniconfig==2.0.0
packaging==24.2
parsimonious==0.9.0
passlib==1.7.4
pillow==11.1.0
pluggy==1.5.0
pycryptodome==3.19.1
pydantic==1.10.21
PyJWT==2.10.1
pymongo==4.11
pytest==7.4.4
python-cartesi==0.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
sqlparse==0.5.3
starlette==0.45.3
tablib==3.8.0
toolz==1.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
whitenoise==6.8.2
